feat: add-review-action-handler impl with i11 tests

+++ src/main/kotlin/com/portfolioy0711/api/data/models/review/ReviewQueryRepository.kt
@@ -2,5 +2,5 @@ package com.portfolioy0711.api.data.models.review
 
 interface ReviewQueryRepository {
     fun checkRecordExistsByReviewId(reviewId: String): Boolean
-    fun findReviewCountsByPlaceId(placeId: String): Int
+    fun findReviewCountsByPlaceId(placeId: String): Long
 }
